requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
firebase-admin
pandas
